22.03 #0

+++ database/migrations/2019_01_09_084100_add_urgency_to_supply_orders_table.php
@@ -26,9 +26,17 @@ class AddUrgencyToSupplyOrdersTable extends Migration
      */
     public function down()
     {
-        /* Проблема, нужна проверка */
-        Schema::table('supply_orders', function (Blueprint $table) {
-            $table->dropColumn('urgency');
-        });
+        // /* Проблема, нужна проверка */
+        // Schema::table('supply_orders', function (Blueprint $table) {
+        //     $table->dropColumn('urgency');
+        // });
+
+        if (Schema::hasColumn('supply_orders', 'urgency'))
+        {
+            Schema::table('supply_orders', function (Blueprint $table)
+            {
+                $table->dropColumn('urgency');
+            });
+        }
     }
 }
